Add libraries necessary to install Sante Fe Swarm
Use these libraries to install Sante Fe Swarm on CENTOS (I used version CentOS Linux release 7.9.2009 (Core)) - do not install gcc or Tk or Tcl with the install - use the libraries here. Install gcc-4.0.0 with the options in the INSTALL file - or Swarm will not compile and run correctly. The SWARM INSTALL NOTES include a lot of things I needed to do to get it to work - you will not need to do these things if you use these libraries.  But the configure command arguments for gcc and other things you will need.  Install gcc (with the options in the INSTALL NOTES) then Tk, (in /usr/local/lib) then tcl then blt then png then swarm, then the examples to test it.